Add completed tasks
Add completed tasks to check if preceding tasks are completed
or not. (Note: completed tasks include completed/canceled
non-joint tasks.)
There is an old logic in workflow.py:
if len(join_task.preceded_by) == len(join_list):
It is used to check if current task is blocked or not, but there
is an exceptional case when removing a subtask before a join
point, original len(join_task.preceded_by)-len(join_list)=1,
after removing a preceding task, the two length will be equal
which result in wrong logic.
This patch add a completed tasks and check if preceding tasks
names are in completed tasks instead of checking the length.

Some minor flake8 issue as well.

JIRA: RHELWF-7410